Move Answer class to a separate file
